Refactor paragraph utils into own module
